requirements.txt
beautifulsoup4==4.9.3
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.21.5
openai==0.18.0
openpyxl==3.0.9
packaging==21.3
pandas-stubs==1.2.0.57
pip==21.2.4
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dotenv==0.20.0
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
setuptools==58.0.4
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
tokenizers==0.10.3
torch==1.11.0
tqdm==4.64.0
transformers==4.16.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
WTForms==2.3.3
zipp==3.7.0
